+++ client.go
@@ -1,91 +1,131 @@
 package gphotos
 
 import (
+	"context"
 	"net/http"
-	"sync"
 
-	"github.com/gphotosuploader/google-photos-api-client-go/v2/internal/cache"
-	"github.com/gphotosuploader/google-photos-api-client-go/v2/internal/log"
-	"github.com/gphotosuploader/google-photos-api-client-go/v2/internal/photoservice"
-	"github.com/gphotosuploader/google-photos-api-client-go/v2/internal/uploader"
+	"github.com/gphotosuploader/google-photos-api-client-go/v2/albums"
+	"github.com/gphotosuploader/google-photos-api-client-go/v2/media_items"
+	"github.com/gphotosuploader/google-photos-api-client-go/v2/uploader"
+	"github.com/gphotosuploader/google-photos-api-client-go/v2/uploader/basic"
+	"github.com/hashicorp/go-retryablehttp"
 )
 
-// Client is a client for uploading a media. photoslibrary does not provide `/v1/uploads` API so we implement here.
+// Client is a Google Photos client with enhanced capabilities.
 type Client struct {
-	// Google Photos client
-	service photoservice.Service
-	// Uploader to upload new files to Google Photos
-	uploader uploader.Uploader
-	// cache to put albums cache
-	cache cache.Cache
-	// mutex to ensure album creation is unique
-	m sync.Mutex
-	// logger to send messages.
-	log log.Logger
+	Albums     albums.AlbumsService
+	MediaItems media_items.MediaItemsService
+	Uploader   uploader.MediaUploader
+}
+
+func (c Client) UploadFileToLibrary(ctx context.Context, filePath string) (media_items.MediaItem, error) {
+	token, err := c.Uploader.UploadFile(ctx, filePath)
+	if err != nil {
+		return media_items.MediaItem{}, err
+	}
+	return c.MediaItems.Create(ctx, media_items.SimpleMediaItem{
+		UploadToken: token,
+		FileName:    filePath,
+	})
+}
+
+func (c Client) UploadFileToAlbum(ctx context.Context, albumId string, filePath string) (media_items.MediaItem, error) {
+	token, err := c.Uploader.UploadFile(ctx, filePath)
+	if err != nil {
+		return media_items.MediaItem{}, err
+	}
+	item := media_items.SimpleMediaItem{
+		UploadToken: token,
+		FileName:    filePath,
+	}
+	return c.MediaItems.CreateToAlbum(ctx, albumId, item)
 }
 
 // NewClient constructs a new gphotos.Client from the provided HTTP client and the given options.
 // The client is an HTTP client used for calling Google Photos. It needs the proper authentication in place.
 //
-// Use WithLogger(), WithCacher(), WithUploader() to customize it.
-func NewClient(httpClient *http.Client, options ...Option) (*Client, error) {
-	var service photoservice.Service
-	var storer uploader.SessionStorer
-	var upldr uploader.Uploader
-	var err error
-	logger := defaultLogger()
-	cacher := defaultCacher()
+// By default it will use a in memory cache for Albums repository and implements retries with Exponential backoff.
+//
+// Use WithUploader(), WithAlbumsService(), WithAlbumsService() to customize it.
+//
+// There is a resumable uploader implemented on uploader.NewResumableUploader().
+func NewClient(authenticatedClient *http.Client, options ...Option) (*Client, error) {
+	client := retryablehttp.NewClient()
+	client.HTTPClient = authenticatedClient
+
+	var albumsService albums.AlbumsService = albums.NewCachedAlbumsService(client.StandardClient())
+
+	var upldr uploader.MediaUploader
+	upldr, err := basic.NewBasicUploader(client.StandardClient())
+	if err != nil {
+		return nil, err
+	}
+
+	var mediaItemsService media_items.MediaItemsService
+	mediaItemsService, err = media_items.NewHttpMediaItemsService(client.StandardClient())
+	if err != nil {
+		return nil, err
+	}
 
 	for _, o := range options {
 		switch o.Name() {
-		case optkeyPhotoService:
-			service = o.Value().(photoservice.Service)
-		case optkeyLogger:
-			logger = o.Value().(log.Logger)
-		case optkeyCacher:
-			cacher = o.Value().(cache.Cache)
-		case optkeySessionStorer:
-			storer = o.Value().(uploader.SessionStorer)
 		case optkeyUploader:
-			upldr = o.Value().(uploader.Uploader)
-		}
-	}
+			upldr = o.Value().(uploader.MediaUploader)
+		case optkeyAlbumsService:
+			albumsService = o.Value().(albums.AlbumsService)
+		case optkeyMediaItemsService:
 
-	// Use GooglePhotosService by default.
-	if service == nil {
-		service, err = photoservice.NewGooglePhotosService(httpClient, WithLogger(logger))
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Use BasicUploader by default, as far as a SessionStorer has not been set.
-	if upldr == nil {
-		if storer == nil {
-			upldr, err = uploader.NewBasicUploader(httpClient, WithLogger(logger))
-			if err != nil {
-				return nil, err
-			}
-		} else {
-			upldr, err = uploader.NewResumableUploader(httpClient, storer, WithLogger(logger))
-			if err != nil {
-				return nil, err
-			}
 		}
 	}
 
 	return &Client{
-		service:  service,
-		uploader: upldr,
-		cache:    cacher,
-		log:      logger,
+		Albums:     albumsService,
+		MediaItems: mediaItemsService,
+		Uploader:   upldr,
 	}, nil
+
 }
 
-func defaultLogger() log.Logger {
-	return &log.DiscardLogger{}
+const (
+	optkeyUploader          = "uploader"
+	optkeyAlbumsService     = "albumService"
+	optkeyMediaItemsService = "mediaItemsService"
+)
+
+// Option represents a configurable parameter.
+type Option interface {
+	Name() string
+	Value() interface{}
 }
 
-func defaultCacher() cache.Cache {
-	return cache.NewCachitaCache()
+type option struct {
+	name  string
+	value interface{}
+}
+
+func (o option) Name() string       { return o.name }
+func (o option) Value() interface{} { return o.value }
+
+// WithUploader configures the Media Uploader.
+func WithUploader(s uploader.MediaUploader) *option {
+	return &option{
+		name:  optkeyUploader,
+		value: s,
+	}
+}
+
+// WithAlbumsService configures the Albums Service.
+func WithAlbumsService(s albums.AlbumsService) *option {
+	return &option{
+		name:  optkeyAlbumsService,
+		value: s,
+	}
+}
+
+// WithMediaItemsService configures the Media Items Service.
+func WithMediaItemsService(s media_items.MediaItemsService) *option {
+	return &option{
+		name:  optkeyMediaItemsService,
+		value: s,
+	}
 }